Simplified UDP handshake process
Removes the need for client to manually send UDP port information to server. Server uses ReceiveFrom() to get the remote EndPoint and stores it for each client. As tested in my Procedural Dungeon Toolkit asset this also works with WAN as the resulting endpoint uses a port on the Public IP that routes to the client's local address.